unify template and type
