+++ src/entities/index.ts
@@ -1,13 +1,11 @@
 import { Notification } from './notification';
 import { NotificationSetting } from './notificationSetting';
-import { NotificationTemplate } from './notificationTemplate';
 import { NotificationType } from './notificationType';
 import { ServiceToken } from './serviceToken';
 import { UserAddress } from './userAddress';
 
 export const entities = [
   Notification,
-  NotificationTemplate,
   NotificationType,
   NotificationSetting,
   ServiceToken,
